actualizacion de saldos despues de transacciones de compra y venta, endpoint de consulta de transacciones, mapa de equivalencia de coind ids

+++ common/src/main/java/org/cryptostream/model/PriceResponse.java
@@ -3,11 +3,12 @@ package org.cryptostream.model;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
 @Builder
 public class PriceResponse {
-    private Map<String, Map<String, Integer>> prices;
+    private Map<String, Map<String, BigDecimal>> prices;
 }
